Users other than the one who made the post, cannot delete, closes #34

+++ BlogAPI/Controllers/PostsController.cs
@@ -229,6 +229,11 @@ namespace BlogAPI.Controllers
                 return NotFound();
             }
 
+            if(post.UserID != postingUser.Id)
+            {
+                return Forbid();
+            }
+
             _context.Posts.Remove(post);
             await _context.SaveChangesAsync();
